- correct temp variable type ( Double ) it was causing mismatch of type for retrofit

+++ openweathermaplibrary/src/main/java/com/psa/openweathermaplibrary/data/source/remote/model/CityWeather.kt
@@ -29,7 +29,7 @@ data class Coord(
 )
 
 data class Main(
-    @SerializedName("temp"       ) var temp      : Int    = 0,
+    @SerializedName("temp"       ) var temp      : Double    = 0.0,
     @SerializedName("feels_like" ) var feelsLike : Double = 0.0,
     @SerializedName("temp_min"   ) var tempMin   : Double = 0.0,
     @SerializedName("temp_max"   ) var tempMax   : Double = 0.0,
